pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "financial-report-agent"
version = "0.1.0"
description = "PDF financial report parsing and trader analysis agent"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [{name = "Your Team"}]
keywords = ["pdf", "finance", "fastapi", "langgraph"]
dependencies = [
  "fastapi>=0.110",
  "uvicorn>=0.24",
  "pydantic>=2.5",
  "langgraph>=0.0.40",
  "typer>=0.9",
  "python-multipart>=0.0.6",
  "python-dotenv>=1.0",
  "pandas>=2.0",
  "structlog>=23.1",
  "pdfplumber>=0.10",
  "PyMuPDF>=1.23",
  "openai>=1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "pytest-asyncio>=0.23",
  "ruff>=0.5",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
